+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Added option to set preferred node id to execute query
 * Set the `pick_first` balancer for short-lived grpc connection inside ydb cluster discovery attempt
 
 ## v3.90.1